requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
appdirs==1.4.4
appnope==0.1.3
astor==0.8.1
asttokens==2.4.1
async-timeout==5.0.1
attrs==24.3.0
audible==0.10.0
audible-cli==0.3.2
backoff==2.2.1
beautifulsoup4==4.12.3
blessed==1.20.0
blinker==1.8.2
cachetools==5.3.2
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.1
contourpy==1.0.7
crewai==0.1.32
cycler==0.11.0
dataclasses-json==0.6.3
debugpy==1.8.0
decorator==5.1.1
distro==1.9.0
editor==1.6.6
eval_type_backport==0.2.0
exceptiongroup==1.2.2
executing==2.0.1
filelock==3.16.1
Flask==3.0.3
fonttools==4.39.4
frozenlist==1.5.0
fsspec==2023.12.2
git-python==1.0.3
gitdb==4.0.11
GitPython==3.1.41
google-api-core==2.15.0
google-api-python-client==2.114.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-search-results==2.1.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
groq==0.11.0
gunicorn==21.2.0
h11==0.14.0
html2image==2.0.4.3
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.20.3
idna==3.10
importlib_metadata==8.5.0
inquirer==3.2.2
ipykernel==6.29.0
ipython==8.20.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.4
langchain==0.1.2
langchain-community==0.0.14
langchain-core==0.1.14
langchain-openai==0.0.2.post1
langgraph==0.0.15
langsmith==0.0.83
litellm==1.20.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
monotonic==1.6
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.24.4
oauthlib==3.2.2
open-interpreter==0.2.0
openai==1.9.0
packaging==23.2
parso==0.8.3
pbkdf2==1.3
pdf2image==1.17.0
pexpect==4.9.0
pillow==10.4.0
pipdeptree==2.7.1
platformdirs==4.1.0
posthog==3.3.3
prompt-toolkit==3.0.43
propcache==0.2.0
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyaes==1.6.1
pyarrow==17.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.0.9
pytesseract==0.3.13
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2022.7
PyYAML==6.0.1
pyzmq==25.1.2
questionary==2.1.0
readchar==4.0.5
regex==2024.11.6
requests==2.32.3
requests-oauthlib==1.3.1
rich==13.9.4
rsa==4.9
runs==1.2.2
scipy==1.10.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.3.2.post1
SQLAlchemy==2.0.25
stack-data==0.6.3
tabulate==0.9.0
tavily-python==0.5.0
tenacity==8.2.3
tiktoken==0.5.2
together==1.3.10
tokenizers==0.15.1
tokentrim==0.1.13
toml==0.10.2
torch==1.13.1
torchvision==0.14.1
tornado==6.4
tqdm==4.67.1
traitlets==5.14.1
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
websocket-client==1.7.0
Werkzeug==3.0.6
wget==3.2
xmod==1.8.1
yarl==1.15.2
zipp==3.20.2
